Исправление багов и улучшения: short→int, seed-данные, метрики, UX админки
- 1.1: short→int для Price/Quantity/Sum/Balance во всех слоях (защита от переполнения)
- 1.4: ImageUrl MaxLength увеличен до 500
- 1.5: UpdateProduct больше не перезаписывает Status при смене количества
- 1.6: Password MaxLength увеличен до 128
- 2.2: UserMapping теперь маппит Branch
- 2.4: Убран дублирующий CreateUser эндпоинт
- 2.5: ProductImageController.Upload обёрнут в транзакцию
- 2.6: OrderService.UpdateOrder убрана избыточная проверка
- 2.7: BatchDelete проверяет существование ID перед удалением
- 3.1: Удалены seed-данные из конфигураций (все данные из БД)
- 3.3: ExceptionHandlerMiddleware возвращает stackTrace в Development
- 3.4: OrderItemService.DeleteOrderItem восстанавливает остаток товара
- 3.5: CancellationTokenSource теперь корректно Dispose перед пересозданием
- 4.1: Сортировка по столбцам (товары/заказы/пользователи), индикатор загрузки формы
- 4.2: Tooltips для статусов, min=0 для баланса, min=1 для цены
- 4.4: Переключатель размера страницы (5/8/16/32)
- 4.5: Метрики считаются по всем данным, исправлено удаление во время итерации

+++ CifraShop.Application/Services/Interfaces/IOrderItemService.cs
@@ -6,7 +6,7 @@ namespace CifraShop.Application.Services.Interfaces
     {
         Task<List<OrderItem>> GetOrderItemsByOrderId(int orderId);
         Task<OrderItem?> GetOrderItemById(int orderItemId);
-        Task<OrderItem> CreateOrderItem(int orderId, int productId, short quantity);
+        Task<OrderItem> CreateOrderItem(int orderId, int productId, int quantity);
         Task UpdateOrderItem(OrderItem orderItemToUpdate);
         Task DeleteOrderItem(OrderItem orderItemToDelete);
     }
