+++ CifraShop.Application/Services/Interfaces/IProductService.cs
@@ -10,10 +10,10 @@ namespace CifraShop.Application.Services.Interfaces
         Task<PagedResponse<Product>> GetProductsPaged(int page, int pageSize, string? search = null, StatusProduct? status = null);
         Task<Product?> GetProductById(int id);
         Task<List<Product>> GetProductsByName(string name);
-        Task<List<Product>> GetProductsByPrice(short price);
-        Task<List<Product>> GetProductsByQuantity(short quantity);
+        Task<List<Product>> GetProductsByPrice(int price);
+        Task<List<Product>> GetProductsByQuantity(int quantity);
         Task<List<Product>> GetProductsByStatus(StatusProduct statusProduct);
-        Task<Product> CreateProduct(string name, string description, short price, short quantity);
+        Task<Product> CreateProduct(string name, string description, int price, int quantity);
         Task UpdateProduct(Product productToUpdate);
         Task DeleteProduct(Product productToDelete);
         Task DeleteRange(List<int> ids);
