+++ CifraShop.Contracts/Requests/OrderItem/CreateOrderItemRequest.cs
@@ -10,7 +10,7 @@ namespace CifraShop.Contracts.Requests.OrderItem
         [Range(1, int.MaxValue, ErrorMessage = "Id товара должен быть больше 0")]
         public int ProductId { get; set; }
 
-        [Range(1, short.MaxValue, ErrorMessage = "Количество должно быть больше 0")]
-        public short Quantity { get; set; }
+        [Range(1, int.MaxValue, ErrorMessage = "Количество должно быть больше 0")]
+        public int Quantity { get; set; }
     }
 }
