+++ CifraShop.Contracts/Requests/OrderItem/UpdateOrderItemRequest.cs
@@ -3,7 +3,7 @@ namespace CifraShop.Contracts.Requests.OrderItem
     public class UpdateOrderItemRequest
     {
         public int Id { get; set; }
-        public short Quantity { get; set; }
-        public short? Price { get; set; }
+        public int Quantity { get; set; }
+        public int? Price { get; set; }
     }
 }
