+++ CifraShop.Contracts/Requests/Products/UpdateProductRequest.cs
@@ -6,8 +6,8 @@ namespace CifraShop.Contracts.Requests.Products
     {
         public string? Name { get; set; }
         public string? Description { get; set; }
-        public short? Price { get; set; }
-        public short? Quantity { get; set; }
+        public int? Price { get; set; }
+        public int? Quantity { get; set; }
         public StatusProduct? Status { get; set; }
     }
 }
